Rename typeForReference -> getTypeForReference
