+++ src/highlight/utils/get-type-for-reference.ts
@@ -3,7 +3,7 @@ import { Reference } from '../types';
 import { isTypeReference } from './is-type-reference';
 import { isFieldReference } from './is-field-reference';
 
-export function typeForReference(schema: GraphQLSchema, reference: Reference): GraphQLNamedType | undefined {
+export function getTypeForReference(schema: GraphQLSchema, reference: Reference): GraphQLNamedType | undefined {
   if (isTypeReference(reference)) {
     return schema.getType(reference) ?? undefined;
   }
